// packages/pieces/community/quickbooks-sandbox/src/lib/types.ts
export interface QuickbooksRef {
    value: string;
    name?: string; // Optional based on context
}

export interface QuickbooksAddress {
    Id?: string;
    Line1?: string;
    City?: string;
    CountrySubDivisionCode?: string;
    PostalCode?: string;
    Lat?: string;
    Long?: string;
}

export interface QuickbooksEmail {
    Address?: string;
}

export interface QuickbooksWebsite {
    URI?: string;
}

export interface QuickbooksPhoneNumber {
    FreeFormNumber?: string;
}

export interface QuickbooksCurrencyRef {
    value: string;
    name?: string;
}

export interface QuickbooksMetaData {
    CreateTime?: string;
    LastUpdatedTime?: string;
}
// ...
export interface QuickbooksVendor {
    Id: string;
    SyncToken?: string;
    DisplayName: string;
    CompanyName?: string;
    PrintOnCheckName?: string;
    Active?: boolean;
    PrimaryPhone?: QuickbooksPhoneNumber;
    PrimaryEmailAddr?: QuickbooksEmail;
    WebAddr?: QuickbooksWebsite;
    BillAddr?: QuickbooksAddress;
    Balance?: number;
    CurrencyRef?: QuickbooksCurrencyRef;
    Vendor1099?: boolean;
    TaxIdentifier?: string;
    TermRef?: QuickbooksRef;
    AcctNum?: string;
    domain?: string;
    sparse?: boolean;
    MetaData?: QuickbooksMetaData;
}
// ...
export interface QuickbooksTerm {
    Id: string;
    Name: string;
    Active?: boolean;
    Type?: string;
    DueDays?: number;
    domain?: string;
    sparse?: boolean;
}